Swap ed25519_dalek crate for ed25519_dalek_bip32.
In order to add support for importing helium-ledger-app seed phrases,
helium-wallet-rs needs to be able to generate BIP-32 public keys. It
seems like the best place to start with that was to add code to do this
in helium-crypto-rs.

This commit is a first step, just swapping out the ed25519_dalek crate
for another which layers in BIP32 support. The existing ed25519_dalek
functionality is still available.

Since this project was depending on specific branch on a Helium fork of
ed25519_dalek, I forked ed25519_dalek_bip32 so it would use the same
Helium version of the underlying ed25519_dalek crate. I don't think you
should actually merge this pointing to my repo, I'm just using those as
an example so a Helium org maintainer can fork ed25519_dalek_bip32 and
see the small change I made to pull in the right branch on Helium's
fork.