Ensure secret_to_vec() returns original entropy. (#39)
The recent change to the ed25519 crate, caused the keypair.secret
member to return something besides the expected original seed
entropy. This created problems for callers of secret_to_vec(),
which expected it to return just the entropy.

The fix is to change
 self.secret.to_vec() -> self.secret.seed().to_vec()

I also added unit tests in the keypair module to assert this
expectation for all key types that implement secret_to_vec().